requirements.txt
tensorflow
pillow
flask
requests
opencv-python
numpy
gunicorn